added code to extract NCEP/CFSR data
Extract daily weather and/or mean monthly climate:
- ExtractSkyDataFromNCEPCFSR_Global
- GriddedDailyWeatherFromNCEPCFSR_Global

Dataset must be made available at
SoilWat_DataSet_External/ExtractSkyDataFromNCEPCFSR_Global

This is also where the C-code ‘cfsr_convert.c’ must be located. The
executable wgrib2 must be available on the system as well.

For details of the setup check the comments in function
‘prepare_NCEPCFSR_extraction’


Former-commit-id: 120579ee21694b78fc36493bd98e5474f968cb8b