Fixing the handling of mean daily output
- commit ‘18826c5’ (Altered SQL commans for daily aggregated output so
it is compatible o…) on branch ‘wrapper_w31’ had changed how the output
for mean daily aggregations was handled: instead of writing it as SQL
statements to a temporary text file (as the output for the overall
aggregations is), this commit added code to add the data directly to
the SQL database. This however must fail with an “database is locked”
error because of multiple processes simultaneously accessing the
database.
- This may work if the database is set to the WAL mode
(http://www.sqlite.org/wal.html). We should test this mode that allows
concurrent read/writes (which is reported to be problematic on NFS
https://groups.google.com/forum/#!topic/batchjobs/15IIm_Nb4YQ)
- For now, I reverted to discussed earlier changes and write the
results again out as SQL statements to the temporary text files. These
are then inserted into the database in the ‘concatenate’ step.

Former-commit-id: 039688d6e384f8e709087ef7c028e8be3f72f2df