New 'do.PriorCalculations' option: 'ExtendSoilDatafileToRequestedSoilLayers'
- new variable defined in part 1 ‘requested_soil_layers’
- if ‘ExtendSoilDatafileToRequestedSoilLayers’ is TRUE, simulation
input for soils is coming from the datafile.soils, and
requested_soil_layers are not existing in datafile.soillayers, then new
soil layers are added and soil characteristics interpolated

- function 'add_layer_to_soil()' from commit
82f22ce302e5735e85b1ab2cbbfbecb9d2de85e7 [formerly dcb7c55709c6f8f95a848e02da6bbeb6190b5659] [formerly 447d3bbf5ada11a1d1eced03140039012e4a945c [formerly 8467228eaf337caf88c84e6bc05e3352e9098528]] updated and gained argument
'method' (interpolate as before or now 'exhaust'). Method 'exhaust'
shares the value of the soil layer to be split in two according to
weights so that the sum remains the same as before.

- options 'ExtendSoilDatafileToRequestedSoilLayers' and
'CalculateBareSoilEvaporationCoefficientsFromSoilTexture' of
'do.PriorCalculations' are now executed even if actions does not
contain 'create', e.g., code is run for data preparation


Former-commit-id: 6983429176a403acc22f60eaa1da4025236bf94a [formerly 95b24d8067aff1858e809d5456479b9ffb6f2b1b]
Former-commit-id: 1ada898a69df2a763a0960e4af732118ef5ea499