New 'do.PriorCalculations' option: 'ExtendSoilDatafileToRequestedSoilLayers'
- new variable defined in part 1 ‘requested_soil_layers’
- if ‘ExtendSoilDatafileToRequestedSoilLayers’ is TRUE, simulation
input for soils is coming from the datafile.soils, and
requested_soil_layers are not existing in datafile.soillayers, then new
soil layers are added and soil characteristics interpolated

- function 'add_layer_to_soil()' from commit
d25272b4fa979d9c8f1911a349939483da187915 updated and gained argument
'method' (interpolate as before or now 'exhaust'). Method 'exhaust'
shares the value of the soil layer to be split in two according to
weights so that the sum remains the same as before.

- options 'ExtendSoilDatafileToRequestedSoilLayers' and
'CalculateBareSoilEvaporationCoefficientsFromSoilTexture' of
'do.PriorCalculations' are now executed even if actions does not
contain 'create', e.g., code is run for data preparation