Rcpp version of germination_wait_times
Profiling the function do_OneSite suggested that
'germination_wait_times' was another major bottleneck: The Rcpp version
of the function is about 270x faster for vectors of length 365 and
12,000x faster for vectors of length 11,000 than the R version. It
appears to scale with length of inputs. The Rcpp version also reduced
the memory footprint by a factor of >> 3080.

Tests:
Both versions (Rcpp and R) pass the unit tests in
‘test_GISSM_germination_wait_times.R’