Rcpp version of germination_wait_times
Profiling the function do_OneSite suggested that
'germination_wait_times' was another major bottleneck: The Rcpp version
of the function is about 270x faster for vectors of length 365 and
12,000x faster for vectors of length 11,000 than the R version. It
appears to scale with length of inputs. The Rcpp version also reduced
the memory footprint by a factor of >> 3080.

Tests:
Both versions (Rcpp and R) pass the unit tests in
‘test_GISSM_germination_wait_times.R’

Former-commit-id: 4670614fb62b8047f537a9dbc637d8ab2d44ead7 [formerly d8525606554c1f41c8f0367b49da75edf3f8e235]
Former-commit-id: 83e94b2d2b02a0da1d154db848be4834fe444f9f