Change functionality of CO2_RCP column in experimental design
Previously, this column set the RCP for a default model (e.g. no climate scenario), in case the user ever wanted something besides the default (85). This has been changed to additionally override the RCPs of scenarios. At the moment there is no functionality to override on a scenario-by-scenario basis. So if a scenario represents RCP45, but the user wanted to use CO2 data for RCP85 (because we lack 45 data), then they would enable this column and set the value to 85. If it is not enabled, the model will use the scenario's RCP.