fix #89 'STATSGO Data: breaks are not unique'
This issue was ultimately being caused because the breaks were not a sequence. I could have just changed the breaks to seq(0, n_cols), but decided to take a more adaptive approach. Now the breaks are a sequence from the minimum value to the maximum value + 1, which should ensure unique breaks.