Parallel framework moved to internal package object
Previously, the cluster object, type of cluster, whether or not a
cluster was set up was handled by the input object 'opt_parallel'.
However, this setup did not allow for a mpi cluster to be up among
different functions that used the cluster. Now all this information is
moved to the internal global package object 'SFSW2_glovars'.