Don't use testthat::skip_on_* outside of test_that() call
skip_on_travis() and skip_on_appveyor() are meant to be used within
test_that() calls. Here, we need to skip preparation code outside of a
test_that call. However, the skip_* functions cause CIs to error out,
even when wrapped in try() statements, with
   - "Error: On Appveyor", respectively
   - "Error: On Travis"
Check values of the ENV variables directly as a work-around.