Fix 'export_parallel_glovars' for MPI
We have to rename functions locally (and export to workers): 'do.call'
(as called by 'mpi.remote.exec'/'mpi.bcast.cmd' of Rmpi v0.6.6) does
not handle 'what' arguments of a character string format "pkg::fun"
because "pkg::fun" is not the name of a function.

But better don't use outdated formals of the function in question!