Enhance 'dailyNRCS_Chambers2014_ResilienceResistance' and 'dailyNRCS_Maestas2016_ResilienceResistance'
- commit 2e766cf0069c917d7444a4b9d69d883f15cf63c4 introduced the
feature that the SMTR variables 'Sregime' and 'Tregime' could contain
partially NAs (if there are not enough 'normal' years for
calculations).
- commit c87a56d299ebe0e9ad445cfd516133fe4dd86d3a accounted for those
NAs in the aggregations 'dailyNRCS_Chambers2014_ResilienceResistance'
and 'dailyNRCS_Maestas2016_ResilienceResistance', but only across the
entire RR table (i.e., either all are NA or all are determined as 1s/0s)

- this commit determines for each RR category separately (instead for
the entire RR table) whether it occurs (1) or not (0), or whether it
cannot be determined (NA).