Fix 'calc.ScenarioWeather': incorrect index in for-loop
The code to 'Double check what time period to choose (the one with the
most overlap to requested years) if multiple netCDF files for selected
combination of scen x var x rip are present' did not update the
variable 'fnc_parts2' which contained the information to subset
'fnc_gcmXscens'. This caused incorrect subsetting of 'fnc_gcmXscens'.

Fix was to iteratively subset 'fnc_parts2' the same way that
'fnc_gcmXscens' was.

Additionally, the subset index 'itemp_remove' was incorrectly
calculated, i.e., 'imax_overlap' was treated as if it were of the same
length as 'itemp2' (now 'ids_fnc') instead of that it contains the
index of 'ids_fnc' to keep.