Fix column type in 'LookupCarbonScenarios.csv'
- The function 'rSFSW2:::SFSW2_read_csv' is reading data with
`iotools::read.csv.raw` if the package `iotools` is present. If so,
then `nrowsClasses` rows are checked for the type/classes of columns.
The default value is 1000.
- The current first 1000 lines of 'LookupCarbonScenarios.csv' contained
0 for the `RCP85` column, i.e., the code interpreted this as logical
type. And values other than 0/1 were turned to NA because they cannot
be represented by this type.
- If this function gets called without `iotools` present, then
`utils::read.csv` is used to read in the file. This function is much
slower, but checks all the rows and was thus able to correctly infer
that we wanted `numeric` and not `logical` type.

- I changed the first line from `0` to `0.0` to tell the code that this
is `numeric` type.