Hack to save `testthat` result to file for post-mortem
Problem: testthat results are not displayed if run with `R CMD check .`; it can be extremely tricky to debug if a test fails only under `R CMD check .`

Hack: Manipulate `testthat:::test_files` so that the relevant object is stored to a file on disk. This object can be loaded post-mortem and analyzed. Interactive example code is provided in the `tests/testthat.R` file.

- addresses https://github.com/r-lib/testthat/issues/795